+++ lib/enum.dart
@@ -1,1 +1,5 @@
 enum ScanType { qr, barcode, defaultMode }
+
+enum ScanFormat { ALL_FORMATS, ONLY_QR_CODE, ONLY_BARCODE }
+
+enum CameraFace { back, front }