+++ lib/flutter_barcode_scanner.dart
@@ -2,6 +2,8 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 
+import 'enum.dart';
+
 /// Please note that this code is reimplementation of https://github.com/AmolGangadhare/flutter_barcode_scanner, since
 /// the package is not updated for quite sometime, I have added here
 /// Scan mode which is either QR code or BARCODE
@@ -25,8 +27,15 @@ class FlutterBarcodeScanner {
   /// Shows a scan line with [lineColor] over a scan window. A flash icon is
   /// displayed if [isShowFlashIcon] is true. The text of the cancel button can
   /// be customized with the [cancelButtonText] string.
-  static Future<String> scanBarcode(String lineColor, String cancelButtonText,
-      bool isShowFlashIcon, ScanMode scanMode, int? delayMillis) async {
+  static Future<String> scanBarcode(
+    String lineColor,
+    String cancelButtonText,
+    bool isShowFlashIcon,
+    ScanMode scanMode,
+    int? delayMillis,
+    String cameraFace,
+    ScanFormat scanFormat,
+  ) async {
     if (cancelButtonText.isEmpty) {
       cancelButtonText = 'Cancel';
     }
@@ -38,7 +47,9 @@ class FlutterBarcodeScanner {
       'isShowFlashIcon': isShowFlashIcon,
       'isContinuousScan': false,
       'scanMode': scanMode.index,
-      'delayMillis': delayMillis ?? 0
+      'delayMillis': delayMillis ?? 0,
+      'cameraFacingText': cameraFace,
+      'scanFormat': scanFormat.name,
     };
 
     /// Get barcode scan result
@@ -55,11 +66,14 @@ class FlutterBarcodeScanner {
   /// be customized with the [cancelButtonText] string. Returns a stream of
   /// detected barcode strings.
   static Stream? getBarcodeStreamReceiver(
-      String lineColor,
-      String cancelButtonText,
-      bool isShowFlashIcon,
-      ScanMode scanMode,
-      int? delayMillis) {
+    String lineColor,
+    String cancelButtonText,
+    bool isShowFlashIcon,
+    ScanMode scanMode,
+    int? delayMillis,
+    String cameraFace,
+    ScanFormat scanFormat,
+  ) {
     if (cancelButtonText.isEmpty) {
       cancelButtonText = 'Cancel';
     }
@@ -71,7 +85,9 @@ class FlutterBarcodeScanner {
       'isShowFlashIcon': isShowFlashIcon,
       'isContinuousScan': true,
       'scanMode': scanMode.index,
-      'delayMillis': delayMillis ?? 0
+      'delayMillis': delayMillis ?? 0,
+      'cameraFacingText': cameraFace,
+      'scanFormat': scanFormat.name,
     };
 
     // Invoke method to open camera, and then create an event channel which will